refactor(ui): extract nav/useShellLayout hook from App.tsx
Moves the XL-viewport media query and the nav-rail wide/narrow cookie
preference (viewportXL, railLabelsWide, toggleRailWidth) into a dedicated
hook. Pure move, no behavior change.
